ADD NAVIGATION WITH HOME & FRIENDS ROUTES/COMPONENTS, MORE FRIENDFORM TO FRIENDLIST COMPONENT
