+++ friends/src/App.js
@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { Route  } from 'react-router-dom';
 import FriendsList from './components/FriendsList';
 import FriendForm from './components/FriendForm';
+import Navigation from './components/Navigation';
+import Home from './components/Home';
 import './App.css';
 
 class App extends Component {
@@ -10,44 +12,33 @@ class App extends Component {
     super();
 
     this.state = {
-      friends: []
+      // friends: []
     }
   }
 
-  componentDidMount() {
-    axios
-      .get('http://localhost:5000/friends')
-      .then(response => {
-          this.setState({friends: response.data})
-      })
-      .catch(error => {
-          console.error('Server Error', error);
-      });
-  }
+  // componentDidMount() {
+  //   axios
+  //     .get('http://localhost:5000/friends')
+  //     .then(response => {
+  //         this.setState({friends: response.data})
+  //     })
+  //     .catch(error => {
+  //         console.error('Server Error', error);
+  //     });
+  // }
 
-  addFriend = friend => {
-    const newFriend = {
-      name: friend.name,
-      age: friend.age,
-      email: friend.email
-    }
-    axios
-      .post('http://localhost:5000/friends', newFriend)
-      .then(response => {
-          this.setState({
-              friends: response.data
-          })
-      })
-      .catch(err =>{
-        console.log(err);
-      })
-  }
+  
 
   render() {
     return (
       <div className="App">
-        <FriendForm addFriend={this.addFriend} /> 
-        <FriendsList friends={this.state.friends} />
+      <Navigation />
+
+      <Route exact path="/" component={Home} />
+      <Route exact path='/friends' component={FriendsList} />
+      
+
+      {/* <FriendsList friends={this.state.friends} /> */}
       </div>
     );
   }
