+++ friends/src/components/Home.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Home = () => <h1></h1>;
+
+export default Home;
